% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/bim_with_target_and_exclusion_regions_removed_fn.R
\name{bim_with_target_and_exclusion_regions_removed_fn}
\alias{bim_with_target_and_exclusion_regions_removed_fn}
\title{Chop out those independent segments of bim that contain 
target markers or exclusion regions}
\usage{
bim_with_target_and_exclusion_regions_removed_fn(params_sampling_set)
}
\arguments{
\item{params_sampling_set, }{and in particular the objects \code{bim},
\code{hotspot}, \code{target_markers}, and \code{exclusion_region}.
See \code{\link{params_sampling_set_fn}}.}
}
\value{
A \code{data.frame} equal to \code{bim} with some lines removed.
The removed lines correspond to those snps that fall in those
independent segments containing \code{target_markers}
or any part of \code{exclusion_region} or any snps that fall inside
any of the hotspots described by \code{hotspot}
}
\description{
For each chromosome, create independent segments of bim
(using hotspots), and throw out any segments that contain any target
snps or any part of an exclusion region.  Areas within hotspots are
also rejected.
}
