% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatars-package.R, R/gatars.R
\docType{package}
\name{gatars}
\alias{gatars}
\alias{gatars-package}
\alias{gatars}
\title{Genetic Association Tests for Arbitrarily Related Subjects}
\usage{
gatars(phenotype, Psi, sampling_set, N_simulated_nulls, weights = NULL)
}
\arguments{
\item{phenotype}{A data.frame containing \code{NNN} rows and at least the two columns
\code{y} and \code{mu}, where \code{y[nnn]} is an indicator for the \code{nnn}-th person
having disease and \code{mu[nnn]} is is a trait probability, the predicted value of 
\code{y[nnn]} based on nongenetic covariates and possibly principal components of ancestry. 
After we saw that principal conponents of ancestry were insignificant, we performed a 
logistic regression of \code{y} on (1) age, (2) membership in family or case/control data, 
and (3) their interaction, and used \code{mu} to be the fitted values of the logistic 
regression.}

\item{Psi}{A matrix with \code{NNN} rows and \code{NNN} columns.  \code{Psi[n1, n2]} is the 
correlation for the genotype at one marker between the \code{n1}-th and \code{n2}-th 
subjects. \code{Psi} can be estimated from known family pedigree structures and/or from
the subjects' genotypes at markers independent of those in the target set. The diagonal 
elements are all unity and two people known to be non-identical-twin full sibs and who
have parents known to be completely unrelated have correlation 0.5.}

\item{sampling_set}{An object of class \code{gatars_sampling_set} produced by 
\code{gatars_sampling_set}. (See \code{\link{gatars_sampling_set}}.)  The object \code{sampling_set} is list containing the objects
\code{params_sampling_set}, \code{report}, and \code{sampling_set}.  The important object
is \code{sampling_set$sampling_set}, a list containing \code{MMM} matrices. When it is time 
for genome resampling, \code{gatars} creates "simulated genotype matrices" by sampling
and cbinding together one column from each of the \code{MMM} matrices.}

\item{N_sim_reps}{An integer equal to the number of "simulated genotype matrices to generate
when calculating the test sizes of the optimal statistics.}
}
\description{
Calculate the p-values of a family of statistics that test the association between a specifed set of genetic markers and a binary or quantitative trait.

(Genetic Association Tests for Arbitrarily Related Subjects) 
Calculate the p-values of a family of statistics that test the association between
a specified set of \code{MMM} makrders and a binary trait.  These statistics include three 
statistics--the squared burden statistic \eqn{Q_B}, the linear kernel (SKAT) statistic
\eqn{Q_S}, a new case-based kernel statistic \eqn{Q_C}--as well as an ensemble of four
statistics \eqn{Q_BS}, \eqn{Q_BC}, \eqn{Q_SC}, \eqn{Q_BSC} that optimize linear combinations
of the three basic statistics.  See \url{http://stanford.edu/~ggong/gatars/} for more details.
}
\details{
\strong{The three functions used in \url{http://stanford.edu/~ggong/gatars/} are:}

\code{gatars_sampling_set}: Create the sampling sets that are used in genome resampling.

\code{gatars}: Uses \code{davies} to calculate the p-values of the basic statistics,
 and genome resampling to calculate the p-values of the optimal statistics.       
  
\code{matrix_image_fn}: A little helper function to display the Psi matrix.

\strong{The datasets used are:}

\code{gatars_example}: A list of the 6 objects you need to provide to run \code{gatars}.

\code{hotspot}: The hotspot data used to create the independent segments that are used by \code{gatars_sampling_set}.

The \code{hotspot} data are from “A Fine-Scale Map of Recombination Rates and Hotspots Across the Human Genome”, Myers et.al., \strong{Science} 
October 2005.
}
\examples{
bim = gatars_example$bim
genotype = gatars_example$genotype
phenotype = gatars_example$phenotype
target_markers = gatars_example$target_markers[3:5]
Psi = gatars_example$Psi
exclusion_region = gatars_example$exclusion_region

NNN = nrow(phenotype)
first_ten = 1:10
last_ten = NNN - (9:0)
matrix_image_fn(Psi[c(first_ten, last_ten), c(first_ten, last_ten)],
                main = "First and last 10 rows and columns of Psi")
library(Matrix)
                
genotype_target_markers = genotype[, target_markers]
list(target_markers = target_markers,
     rank = as.numeric(rankMatrix( genotype_target_markers)))
     
set.seed(2)
epsilon = 0.01
sampling_set = gatars_sampling_set(
    bim, epsilon, exclusion_region,
    genotype, hotspot, target_markers)
print(sampling_set)

N_simulated_nulls = 10
gatars(phenotype, Psi, sampling_set, N_simulated_nulls)
}
