% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_segment_fn.R
\name{independent_segment_fn}
\alias{independent_segment_fn}
\title{Chop a chromosome into independent segments}
\usage{
independent_segment_fn(bim, chromosome, hotspot)
}
\arguments{
\item{bim}{The bim object described in \code{link{gatars_sampling_set}}.}

\item{chromosome}{An integer equal to the chromosome number (1 through 22)
of the chromosome you want to chop up.}

\item{hotspot}{The hotspot object described in \code{link{gatars_sampling_set}}.}
}
\value{
A list of \code{data.frame}'s, one \code{data.frame} for each
indpendent segment.  Each \code{data.frame} contains one or few rows from
\code{bim}, and has columns \code{chromosomse}, \code{snp}, and \code{bp}.
}
\description{
(a) Subset \code{bim} to just those
inside \code{chromosome}, (b) discard the head and tail (those bp falling 
outside the first and last hotspots of \code{chromosome}),
and (c) using \code{hotspot} cut up the remaining \code{bim} 
into independent segments.
}
\examples{
bim = gatars_example$bim
chromosome = 21 # I chose 21 because in our example it is the shortest
table(bim$chromosome)
independent_segment_fn_answer = independent_segment_fn(bim, chromosome, hotspot)
length(independent_segment_fn_answer)
head(independent_segment_fn_answer)

}
