% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_optimized_fn.R
\name{p_value_optimized_fn}
\alias{p_value_optimized_fn}
\title{Estimate the p-values of the optimized statistics}
\usage{
p_value_optimized_fn(adaptive_conf_level, calculate_optimized, MMM,
  N_simulated_nulls_interval, N_simulated_nulls_limit, Phi, sampling_set,
  test_size, theta, WWW, x_observed, y_1, y_2)
}
