% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_fn.R
\name{params_fn}
\alias{params_fn}
\title{The parameters that run the innards of \code{uno_experimento_fn}}
\usage{
params_fn(alpha_uni_N_increments, params_sampling_set, phenotype, Psi,
  sampling_set, N_simulated_nulls, weights = NULL)
}
\arguments{
\item{bim}{A \code{data.frame} containing the three columns
\code{chromosome}, \code{snp}, \code{bp}, and containing \code{LLL} rows 
corresponding to the \code{MMM} target markers and those used to build
the sampling sets. The \code{lll}-th row of \code{bim} summarizes the 
\code{lll}-th marker and corresponds to the \code{lll}-th column of 
\code{genotype} (described below).  The column labeled \code{chromosome} 
contains integers between \code{1} and \code{22} (other integers may be 
included, but only chromosomes \code{1} through \code{22} will be used),
the \code{snp} column contains character strings that name the marker,
and the \code{bp} column contains integers giving the markers' bp
positions.  Because the \code{gatars} dat set \code{hotspot} (described
below) is given in Build hg38/GRCh38, \code{bp} must also be expressed in
Build hg38/GRCh38.) The object \code{bim} could be the 
\code{.map} or \code{.bim} file from plink.}
}
\description{
This is my attempt to make the variables that run
\code{uno_experimento_fn} global but safer. Massage the inputs
and output the variables that will be neccessary for the innards
of \code{uno_experimento_fn}.
}
