% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_sampling_set_fn.R
\name{params_sampling_set_fn}
\alias{params_sampling_set_fn}
\title{The parameters that run the innards of gatars_sampling_set}
\usage{
params_sampling_set_fn(bim, epsilon, exclusion_region, genotype, hotspot,
  target_markers)
}
\arguments{
\item{bim}{A \code{data.frame} containing the three columns
\code{chromosome}, \code{snp}, \code{bp}, and containing \code{LLL} rows 
corresponding to the \code{MMM} target markers and those used to build
the sampling sets. The \code{lll}-th row of \code{bim} summarizes the 
\code{lll}-th marker and corresponds to the \code{lll}-th column of 
\code{genotype} (described below).  The column labeled \code{chromosome} 
contains integers between \code{1} and \code{22} (other integers may be 
included, but only chromosomes \code{1} through \code{22} will be used),
the \code{snp} column contains character strings that name the marker,
and the \code{bp} column contains integers giving the markers' bp
positions.  Because the \code{gatars} dat set \code{hotspot} (described
below) is given in Build hg38/GRCh38, \code{bp} must also be expressed in
Build hg38/GRCh38.) The object \code{bim} could be the 
\code{.map} or \code{.bim} file from plink.}

\item{epsilon}{A positive small real number used to determine if a marker
can be included in a sampling set.
When creating the \code{mmm}-th sampling set for a target marker with 
minor allele frequency \code{pi[mmm]} only those markers whose minor allele 
frequencies fall
within the interval \code{pi[mmm] * [1 - epsilon, 1 + epsilon]}
can be included in the sampling set. In our simulations, we set
\code{epsilon} equal to 0.01.}

\item{exclusion_region}{A \code{data.frame} with one or several rows of the three
columns \code{chromosome}, \code{start}, and \code{end}.  Each row of
\code{exclusion_region} reflects one contiguous genomic region known to be
trait-associated and therefore a region used by \code{gatars} when creating the
sampling sets.  The column \code{chromosome} is an integer between \code{1} and
\code{22} identifying the autosomal choromosome containing the region, and
\code{start} and \code{end} describe its starting and ending positions \code{bp}.
If the region consists of a single marker, then \code{start} and \code{end}
are both equal to the position of this marker.  \code{start} and \code{end}
must be expressed in Build hg38/GRCh38.}

\item{genotype}{A matrix with \code{NNN} rows and \code{LLL} columns, whose 
\code{(nnn,lll)}-th element records the \code{nnn}-th subject's coded genotype
at the \code{lll}-th of \code{LLL} markers.  The \code{lll}-th column of 
\code{genotype} corresponds to the \code{lll}-th row of \code{bim}.  The object
\code{genotype} could be obtained by reading in the \code{.bed} file from
plink.  (Here we distinguish \code{genotype} the \code{NNN} x \code{LLL} matrix
whose columns correspond to the \code{MMM} target markers AND the
additional \code{LLL - MMM} markers that are candidates for membership
into sampling sets) from the \code{NNN}
x \code{MMM} matrix G, described in the manuscript, containing just the \code{MMM}
target markers).}

\item{hotspot}{\pkg{gatars} provides this data set for your convenience.
This data.frame must contain (at least) the columns \code{chromosome},
\code{start},  and \code{end}; \code{chromosome} gives the chromosome number
(an integer in \code{1, ..., 22}), and \code{start}, and \code{end} give the extent of the 
recombination hotspot (integers in bp units using Build hg38/GRCH38).}

\item{target_markers}{A character vector of length \code{MMM} that is a subset 
of the column \code{bim$snp}. This vector identifies the target markers.}
}
\value{
A list containing the following 11 objects which are used in the innards of 
\code{gatars_sampling_set}
\itemize{
\item{\code{bim}: } {
The same as the input \code{bim}.
}
\item{\code{genotype}: } {
The same as the input \code{genotype}.
}
\item{\code{e_g_target}: }{
A matrix with `NNN` rows and `MMM` columns.  All rows are the same.  
The \code{mmm}-th column contains the mean of the \code{mmm}-th
column of \code{g_target}.
}
\item{\code{epsilon}: } {
The same as the input \code{epsilon}.
}
\item{\code{exclusion_region}: } {
The same as the input \code{exclusion region}.
}
\item{\code{g_target}: } {
A submatrix of \code{genotype} containing just the columns
specified by \code{target_markers}. This matrix is equal to 
"the genotype matrix" described in the manuscript.
}
\item{\code{hotspot}: } {
The same as the input \code{hotspot}.
}
\item{\code{MMM}: } {
The length of \code{target_markers} and the number of columns of 
\code{g_target}
}
\item{\code{NNN}: } {
The number of rows of \code{genotype} and
\code{g_target}
}
\item{\code{p_target}: } {
A vector of length \code{MMM} whose \code{mmm}-th
element is the mean of the \code{mmm}-th column column of
\code{g_target} divided by two. The website 
\url{http://stanford.edu/~ggong/gatars/} refers to this
vector as \code{pi}.
}
\item{\code{target_markers}: } {
The same as the input \code{target_markers}.
}
}
}
\description{
Most of the parameters used in gatars_sampling_set are 
inputs that are passed through.  \code{g_target}, \code{p_target},
\code{e_g_target}, \code{MMM}, and \code{NNN} are calculated from
the inputs.
}
\examples{
bim = gatars_example$bim
epsilon = 0.01
exclusion_region = gatars_example$exclusion_region
genotype = gatars_example$genotype
str(hotspot)
target_markers = gatars_example$target_markers[3:5]
params_sampling_set = params_sampling_set_fn(
  bim, epsilon, exclusion_region,
  genotype, hotspot, target_markers)
names(params_sampling_set)
}
