% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_set_fn.R
\name{sampling_set_fn}
\alias{sampling_set_fn}
\title{Create sampling sets}
\usage{
sampling_set_fn(params_sampling_set)
}
\arguments{
\item{params_sampling_set, }{and in particular the objects 
\code{epsilon_on_log_scale}, \code{MMM}, and \code{p_target}.
See \code{\link{params_sampling_set_fn}}.}
}
\value{
A \code{list} containing the following two objects
\itemize{
\item{\code{report}: } {
A \code{data.frame} containing \code{MMM} rows and the columns
\code{min}, \code{p_target}, \code{max}, and \code{set_size}.
\code{min}/\code{max} contains the smallest/largest maf among
the columns in the \code{mmm}-th sampling set, \code{p_target}
the maf of the \code{mmm}-th target marker, and \code{set_size}
the number of columns in the \code{mmm}-th sampling set.
}
\item{\code{sampling_set}: }{
A list of \code{MMM} matrices, one matrix for each target snp.
The \code{mmm}-th matrix is the sampling set for the \code{mmm}-th
target snp and has \code{NNN} rows and up to 1000 columns, each 
column containing a column from \code{genotype}.  These columns
do not intersect with any of the target snps or exclusion regions
and the mafs of the columns in \code{mmm}-th sampling set match
the maf of the \code{mmm}-th target snp.
}
}
}
\description{
For each target marker, gatars requires a sampling set,
a collection of columns from \code{genotype}.  Each sampling set
must contain snps that (1) do not intersect with any of the target snps or 
exclusion regions, and (2) have mafs (minor allele frequencies)
that match the maf of its
corresponding target snp.
}
\examples{
library(gatars)
bim = alternative_example$bim
epsilon_on_log_scale = 0.02
exclusion_region = alternative_example$exclusion_region
genotype = alternative_example$genotype
target_markers = alternative_example$target_markers[3:5]
set.seed(1)
params_sampling_set = params_sampling_set_fn(
  bim, epsilon_on_log_scale, exclusion_region,
  genotype, hotspot, target_markers)
names(params_sampling_set)
sampling_set_fn_answer = sampling_set_fn(params_sampling_set)
names(sampling_set_fn_answer)

}
