% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatars_sampling_set.R
\name{gatars_sampling_set}
\alias{gatars_sampling_set}
\title{Create sampling sets for genome resampling}
\usage{
gatars_sampling_set(bim, epsilon, exclusion_region, genotype, hotspot,
  target_markers)
}
\arguments{
\item{bim}{A \code{data.frame} containing the three columns
\code{chromosome}, \code{snp}, \code{bp}, and containing \code{LLL} rows 
corresponding to the \code{MMM} target markers and those used to build
the sampling sets. The \code{lll}-th row of \code{bim} summarizes the 
\code{lll}-th marker and corresponds to the \code{lll}-th column of 
\code{genotype} (described below).  The column labeled \code{chromosome} 
contains integers between \code{1} and \code{22} (other integers may be 
included, but only chromosomes \code{1} through \code{22} will be used),
the \code{snp} column contains character strings that name the marker,
and the \code{bp} column contains integers giving the markers' bp
positions.  Because the \code{gatars} dat set \code{hotspot} (described
below) is given in Build hg38/GRCh38, \code{bp} must also be expressed in
Build hg38/GRCh38.) The object \code{bim} could be the 
\code{.map} or \code{.bim} file from plink.}

\item{epsilon}{A positive small real number used to determine if a marker
can be included in a sampling set.
When creating the \code{mmm}-th sampling set for a target marker with 
minor allele frequency \code{pi[mmm]} only those markers whose minor allele 
frequencies fall
within the interval \code{pi[mmm] * [1 - epsilon, 1 + epsilon]}
can be included in the sampling set. In our simulations, we set
\code{epsilon} equal to 0.01.}

\item{exclusion_region}{A \code{data.frame} with one or several rows of the three
columns \code{chromosome}, \code{start}, and \code{end}.  Each row of
\code{exclusion_region} reflects one contiguous genomic region known to be
trait-associated and therefore a region used by \code{gatars} when creating the
sampling sets.  The column \code{chromosome} is an integer between \code{1} and
\code{22} identifying the autosomal choromosome containing the region, and
\code{start} and \code{end} describe its starting and ending positions \code{bp}.
If the region consists of a single marker, then \code{start} and \code{end}
are both equal to the position of this marker.  \code{start} and \code{end}
must be expressed in Build hg38/GRCh38.}

\item{genotype}{A matrix with \code{NNN} rows and \code{LLL} columns, whose 
\code{(nnn,lll)}-th element records the \code{nnn}-th subject's coded genotype
at the \code{lll}-th of \code{LLL} markers.  The \code{lll}-th column of 
\code{genotype} corresponds to the \code{lll}-th row of \code{bim}.  The object
\code{genotype} could be obtained by reading in the \code{.bed} file from
plink.  (Here we distinguish \code{genotype} the \code{NNN} x \code{LLL} matrix
whose columns correspond to the \code{MMM} target markers AND the
additional \code{LLL - MMM} sampling set marker candidates) from the \code{NNN}
x \code{MMM} matrix G, described in the manuscript, containing just the \code{MMM}
target markers).}

\item{hotspot}{\pkg{gatars} provides this data set for your convenience.
This data.frame must contain (at least) the columns \code{chromosome},
\code{start},  and \code{end}; \code{chromosome} gives the chromosome number
(\code{1, ..., 22}), and \code{start}, and \code{end} give the extent of the 
recombination hotspot (integers in bp units using Build hg38/GRCH38).}

\item{target_markers}{A character vector of length \code{MMM} that is a subset 
of the column \code{bim$snp}. This vector identifies the target markers.}
}
\value{
A list containing the following objects
\itemize{
\item{\code{params_sampling_set}: } {
A list containing objects that will be useful in the calculations of 
the \code{gatars_test_size} function.
The result of a call to \code{params_sampling_set}.  
See \code{\link{params_sampling_set_fn}}.
}
\item{\code{report}: } {
A \code{data.frame} containing \code{MMM} rows and the columns
\code{min}, \code{pi}, \code{max}, and \code{set_size}.
\code{min}/\code{max} contains the smallest/largest 
minor allele frequency among
the columns in the \code{mmm}-th sampling set, \code{pi}
the minor allele frequency of the \code{mmm}-th target marker, and \code{set_size}
the number of columns in the \code{mmm}-th sampling set.
}
\item{\code{sampling_set}: }{
A list of \code{MMM} matrices, one matrix for each target marker.
The \code{mmm}-th matrix is the sampling set for the \code{mmm}-th
target marker and has \code{NNN} rows and up to code{1000} columns, each 
column containing a column from \code{genotype}.  These columns
do not intersect with any of the target markers or exclusion regions
and the minor allele frequencies of the columns in \code{mmm}-th sampling set match
the minor allele frequency of the \code{mmm}-th target marker.
}
}
}
\description{
Testing the null hypothesis for the optimized statistics requires
sampling sets which are sets of markers located throughout the autosomal
genome satisfying some crucial conditions. \code{gatars_sampling_set} 
creates \code{MMM} such sampling sets,  one for 
for each target marker, with the sampling set for the \code{mmm}-th target 
marker containing markers all of whose minor allele frequencies match that 
of the \code{mmm}-th target marker and such that markers in the sampling 
sets are independent of the target markers and of any markers known to be 
associated with the trait.
}
\details{
Each column in \code{genotype} corresponds to a row in \code{bim}, and both 
correspond to a marker. Assuming that markers within two consecutive hotspots 
are independent of those within any other two consecutive hotspots, the recombination 
hotspots divide the autosomal genome contained in the columns of \code{genotype}
into independent segments.  Remove all segments containing any target markers or 
any markers defined by the \code{exclusion_region} data set. On the markers in the 
remaining segments, calculate the empirical minor allele frequencies and say that 
a marker's minor allele frequency  matches the minor allele frequency \code{pi[mmm]} 
of the \code{mmm}-th target marker if it falls in the closed interval
\code{pi[mmm] * [1 - epsilon, 1 + epsilon]}.  (If the number of markers satisfying the 
matching requirement exceeds \code{1000}, \code{gatars} randomly chooses \code{1000}.)
}
\examples{
bim = gatars_example$bim
epsilon = 0.01
exclusion_region = gatars_example$exclusion_region
genotype = gatars_example$genotype
str(hotspot)
target_markers = gatars_example$target_markers[3:5]
sampling_set = gatars_sampling_set(
    bim, epsilon, exclusion_region,
    genotype, hotspot, target_markers)
print(sampling_set)
names(sampling_set)
names(sampling_set$params_sampling_set)
sampling_set$report
str(sampling_set$sampling_set)
}
