% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatars_test_size.R
\name{gatars_test_size}
\alias{gatars_test_size}
\title{Genetic Association Tests for Arbitrarily Related Subjects}
\usage{
gatars_test_size(phenotype, Psi, sampling_set, N_simulated_nulls,
  weights = NULL)
}
\arguments{
\item{phenotype}{A data.frame containing \code{NNN} rows and at least the two columns
\code{y} and \code{mu}.  \code{y[nnn]} is the binary or quantitative trait of the 
\code{nnn}-th person.  If the trait is binary, possible values are \code{0} for not affected
and \code{1} for affected; if the trait is quantitative, possible values are real numbers.
\code{mu[nnn]}, a real number, is  a trait prediction of 
\code{y[nnn]}.  (For example, \code{mu[nnn]} can be the mean of the \code{NNN} \code{y[nnn]}
values or the value predicted from a logistic or linear regression of \code{y[nnn]}
on nongenetic covariates and/or on principal components of ancestry.)}

\item{Psi}{A real-valued matrix with \code{NNN} rows and \code{NNN} columns.
\code{Psi[n1, n2]} is the 
correlation for the genotype at one marker between the \code{n1}-th and \code{n2}-th 
subjects. \code{Psi} can be estimated from known family pedigree structures and/or from
the subjects' genotypes at markers independent of those in the target set. The diagonal 
elements are all unity and two people known to be non-identical-twin full sibs and who
have parents known to be completely unrelated have correlation 0.5.}

\item{sampling_set}{An object of class \code{gatars_sampling_set} produced by 
\code{gatars_sampling_set}. (See \code{\link{gatars_sampling_set}}.)  
The object \code{sampling_set} is list containing the objects
\code{params_sampling_set}, \code{report}, and \code{sampling_set}.  The important object
is \code{sampling_set$sampling_set}, a list containing \code{MMM} matrices. 
To run genome resampling, \code{gatars_test_size}
creates "simulated genotype matrices" by sampling
and binding together one column from each of the \code{MMM} matrices.}

\item{N_simulated_nulls}{An integer equal to the number of simulated genotype matrices to generate
when calculating the test sizes of the optimized statistics.}
}
\description{
Test the association between a prespecified set of genetic markers, called target markers, 
and a binary or quantitative trait using subjects with any genealogical relationship
by calculating the p-values of the follwoing seven statistics: the three basic
statistics given by the squared burden \code{Q_B}, the linear kernel (SKAT) statistic
\eqn{Q_S}, and a new trait-based kernel statistic \eqn{Q_T}--as well as an ensemble of four
statistics \eqn{Q_BS}, \eqn{Q_BT}, \eqn{Q_ST}, \eqn{Q_BST} that optimize linear combinations
of the three basic statistics.  See \url{http://stanford.edu/~ggong/gatars/} for more details.
}
