% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_image_fn.R
\name{matrix_image_fn}
\alias{matrix_image_fn}
\title{Display a correlation type matrix}
\usage{
matrix_image_fn(the_matrix, main = "", legend_name = "",
  vertical_line_after = NULL, horizontal_line_after = NULL, size = 3,
  linetype = 3, xlab = "", ylab = "", row_ticks = 1:nrow(the_matrix),
  column_ticks = 1:ncol(the_matrix), legend_position = "right")
}
\arguments{
\item{the_matrix}{The matrix you would like to display.}

\item{main}{The title if you would like one.}
}
\description{
\code{matrix_image_fn} uses \code{geom_tile} from \code{ggplot2} to display a 
correlation type image in the orientation that I think a mathematician would
expect.
}
\examples{
set.seed(1)
NNN = 30
epsilon = .8
x = rnorm(NNN, 0, 1) 
x1 = x + rnorm(NNN, 0, epsilon)
x2 = x + rnorm(NNN, 0, epsilon)
x3 = x + rnorm(NNN, 0, epsilon)
independent_columns = matrix(rnorm(5 * NNN, 0, 1), ncol = 5)
data = cbind(x1, x2, x3, independent_columns)
matrix_image_fn(cor(data), "The correlation of some multivariate normal data")
}
