% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_sampling_set_fn.R
\name{params_sampling_set_fn}
\alias{params_sampling_set_fn}
\title{The parameters that run the innards of gatars_sampling_set}
\usage{
params_sampling_set_fn(bim, epsilon_on_log_scale, exclusion_region, genotype,
  hotspot, target_markers)
}
\arguments{
\item{bim}{A data.frame containing \code{LLL} rows corresponding to
(a very large number of) \code{LLL} markers including the target markers
and candidates for the sampling sets. The columns of \code{bim} 
must contain  \code{chromosome}, \code{snp}, and \code{bp}, giving
respectively,
for each marker its chromosome number, its name, and its base-pair
position (in bp units).  The object \code{bim} could be the 
\code{.map} or \code{.bim} file from plink. 
The \code{lll}-th row of \code{bim} corresponds to the \code{lll}-th column of 
\code{genotype}.  
\code{bp} must be expressed in the same build as \code{hotspot}, and if you 
are using \code{hotspot} provided by \code{gatars} this build is hg38/GRCh38.}

\item{epsilon_on_log_scale}{A positive real number used to parametrize the matching.
For the \code{mmm}-th sampling set of a target marker with minor allele
frequency \code{pi[mmm]}, only those markers whose minor allele frequencies fall
in the interval \code{pi[mmm] * [1 - epsilon_on_log_scale, 1 + epsilon_on_log_scale]}
are candidates for its the sampling set. In our simulations we set
\code{epsilon_on_log_scale} equal to 0.02.}

\item{exclusion_region}{A \code{data.frame} with one or several rows of the three
columns \code{chromosome}, \code{start}, and \code{end}.  Each row of this
\code{data.frame} should reflect one contiguous genomic region known to be associated
with the binary or continuous trait.  
The column \code{chromosome} is an integer between
\code{1} and \code{22} and names in which chromosome the region lies, and \code{start}
and \code{end} describe the starting and ending positions (in bp units) of the contiguous
region.
If the region consists of a single marker, then \code{start} and \code{end} should
both be set equal to the position of this marker.  \code{start} and \code{end} must be 
expressed in Build hg38/GRCh38 for the same reason the column \code{bp} in
\code{bim} must be expressed in Build hg38/GRCh38.}

\item{genotype}{A matrix with \code{NNN} rows and \code{LLL} columns, whose 
\code{(nnn,lll)}-th element records either the number (0, 1 or 2) of minor alleles 
of snp \code{lll} found in the \code{nnn}-th subject or an indicator (0 or 1) 
for the \code{nnn}-th person carrying at least one minor allele of snp \code{lll}. 
The \code{lll}-th column of \code{genotype} corresponds to the \code{lll}-th row of 
\code{bim}. The object \code{genotype} could be gotten by reading in the \code{.bed} file 
from plink after massaging genotype information into either dosage or carriage. 
(Distinguish the object \code{genotype} here, containing target markers AND 
sampling set snps from the “genotype matrix” denoted by \eqn{G} in the manuscript, 
the matrix containing just the target markers.)}

\item{hotspot}{The \pkg{gatars} package provides this data set for your 
convenience.  This data.frame must contain (at least) the columns \code{chromosome},
\code{start}, \code{center}, and \code{end};
 \code{chromosome} describes the chromosome number
(\code{1, ..., 22}), and \code{start}, \code{center}, \code{end}
 describe the location of hotspots 
(in bp units using Build hg38/GRCH38).}

\item{target_markers}{A character vector of length \code{MMM} that is a subset 
of the column \code{bim$snp}. This vector names the target markers.}
}
\value{
A list containing the following 11 objects which are used in the innards of 
\code{gatars_sampling_set}
\itemize{
\item{\code{bim}: } {
The same as the input \code{bim}.
}
\item{\code{genotype}: } {
The same as the input \code{genotype}.
}
\item{\code{e_g_target}: }{
A matrix with `NNN` rows and `MMM` columns.  All rows are the same.  
The \code{mmm}-th column contains the mean of the \code{mmm}-th
column of \code{g_target}.
}
\item{\code{epsilon_on_log_scale}: } {
The same as the input \code{epsilon_on_log_scale}.
}
\item{\code{exclusion_region}: } {
The same as the input \code{exclusion region}.
}
\item{\code{g_target}: } {
A submatrix of \code{genotype} containing just the columns
specified by \code{target_markers}.
}
\item{\code{hotspot}: } {
The same as the input \code{hotspot}.
}
\item{\code{MMM}: } {
The length of \code{target_markers} and the number of columns of 
\code{g_target}
}
\item{\code{NNN}: } {
The number of rows of \code{genotype} and
\code{g_target}
}
\item{\code{p_target}: } {
A vector of length \code{MMM} whose \code{mmm}-th
element is the mean of the \code{mmm}-th column column of
\code{g_target} divided by two.
}
\item{\code{target_markers}: } {
The same as the input \code{target_markers}.
}
}
}
\description{
Most of the parameters used in gatars_sampling_set are 
inputs that are passed through.  \code{g_target}, \code{p_target},
\code{e_g_target}, \code{MMM}, and \code{NNN} are calculated from
the inputs.
}
\examples{
bim = gatars_example$bim
epsilon_on_log_scale = 0.02
exclusion_region = gatars_example$exclusion_region
genotype = gatars_example$genotype
str(hotspot)
target_markers = gatars_example$target_markers[3:5]
params_sampling_set = params_sampling_set_fn(
  bim, epsilon_on_log_scale, exclusion_region,
  genotype, hotspot, target_markers)
names(params_sampling_set)
}
